Faucet.c - AOE attack more or less done, each droplet hit reduces opacity by one,
When adding to upgrades, it should be considered that the reduction is quite strong, should be implemented only in later stages.

+++ Game_Project/dirt.h
@@ -9,7 +9,7 @@ void reduce_all_dirt_opacity(int);
 typedef struct dirt {
     float positionX;
     float positionY;
-    int opacity;
+    float opacity;
 } dirt;
 
 extern dirt dirtList[];
